+++ src/m3ninx/doc/doc_mock.go
@@ -27,6 +27,8 @@ package doc
 import (
 	"reflect"
 
+	"github.com/m3db/m3/src/x/time"
+
 	"github.com/golang/mock/gomock"
 )
 
@@ -279,4 +281,132 @@ func (m *MockQueryDocIterator) Next() bool {
 func (mr *MockQueryDocIteratorMockRecorder) Next() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Next", reflect.TypeOf((*MockQueryDocIterator)(nil).Next))
+}
+
+// MockOnIndexSeries is a mock of OnIndexSeries interface.
+type MockOnIndexSeries struct {
+	ctrl     *gomock.Controller
+	recorder *MockOnIndexSeriesMockRecorder
+}
+
+// MockOnIndexSeriesMockRecorder is the mock recorder for MockOnIndexSeries.
+type MockOnIndexSeriesMockRecorder struct {
+	mock *MockOnIndexSeries
+}
+
+// NewMockOnIndexSeries creates a new mock instance.
+func NewMockOnIndexSeries(ctrl *gomock.Controller) *MockOnIndexSeries {
+	mock := &MockOnIndexSeries{ctrl: ctrl}
+	mock.recorder = &MockOnIndexSeriesMockRecorder{mock}
+	return mock
+}
+
+// EXPECT returns an object that allows the caller to indicate expected use.
+func (m *MockOnIndexSeries) EXPECT() *MockOnIndexSeriesMockRecorder {
+	return m.recorder
+}
+
+// DecrementReaderWriterCount mocks base method.
+func (m *MockOnIndexSeries) DecrementReaderWriterCount() {
+	m.ctrl.T.Helper()
+	m.ctrl.Call(m, "DecrementReaderWriterCount")
+}
+
+// DecrementReaderWriterCount indicates an expected call of DecrementReaderWriterCount.
+func (mr *MockOnIndexSeriesMockRecorder) DecrementReaderWriterCount() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DecrementReaderWriterCount", reflect.TypeOf((*MockOnIndexSeries)(nil).DecrementReaderWriterCount))
+}
+
+// IfAlreadyIndexedMarkIndexSuccessAndFinalize mocks base method.
+func (m *MockOnIndexSeries) IfAlreadyIndexedMarkIndexSuccessAndFinalize(blockStart time.UnixNano) bool {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "IfAlreadyIndexedMarkIndexSuccessAndFinalize", blockStart)
+	ret0, _ := ret[0].(bool)
+	return ret0
+}
+
+// IfAlreadyIndexedMarkIndexSuccessAndFinalize indicates an expected call of IfAlreadyIndexedMarkIndexSuccessAndFinalize.
+func (mr *MockOnIndexSeriesMockRecorder) IfAlreadyIndexedMarkIndexSuccessAndFinalize(blockStart interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IfAlreadyIndexedMarkIndexSuccessAndFinalize", reflect.TypeOf((*MockOnIndexSeries)(nil).IfAlreadyIndexedMarkIndexSuccessAndFinalize), blockStart)
+}
+
+// IndexedForBlockStart mocks base method.
+func (m *MockOnIndexSeries) IndexedForBlockStart(indexBlockStart time.UnixNano) bool {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "IndexedForBlockStart", indexBlockStart)
+	ret0, _ := ret[0].(bool)
+	return ret0
+}
+
+// IndexedForBlockStart indicates an expected call of IndexedForBlockStart.
+func (mr *MockOnIndexSeriesMockRecorder) IndexedForBlockStart(indexBlockStart interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IndexedForBlockStart", reflect.TypeOf((*MockOnIndexSeries)(nil).IndexedForBlockStart), indexBlockStart)
+}
+
+// NeedsIndexUpdate mocks base method.
+func (m *MockOnIndexSeries) NeedsIndexUpdate(indexBlockStartForWrite time.UnixNano) bool {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "NeedsIndexUpdate", indexBlockStartForWrite)
+	ret0, _ := ret[0].(bool)
+	return ret0
+}
+
+// NeedsIndexUpdate indicates an expected call of NeedsIndexUpdate.
+func (mr *MockOnIndexSeriesMockRecorder) NeedsIndexUpdate(indexBlockStartForWrite interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NeedsIndexUpdate", reflect.TypeOf((*MockOnIndexSeries)(nil).NeedsIndexUpdate), indexBlockStartForWrite)
+}
+
+// OnIndexFinalize mocks base method.
+func (m *MockOnIndexSeries) OnIndexFinalize(blockStart time.UnixNano) {
+	m.ctrl.T.Helper()
+	m.ctrl.Call(m, "OnIndexFinalize", blockStart)
+}
+
+// OnIndexFinalize indicates an expected call of OnIndexFinalize.
+func (mr *MockOnIndexSeriesMockRecorder) OnIndexFinalize(blockStart interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnIndexFinalize", reflect.TypeOf((*MockOnIndexSeries)(nil).OnIndexFinalize), blockStart)
+}
+
+// OnIndexPrepare mocks base method.
+func (m *MockOnIndexSeries) OnIndexPrepare(blockStart time.UnixNano) {
+	m.ctrl.T.Helper()
+	m.ctrl.Call(m, "OnIndexPrepare", blockStart)
+}
+
+// OnIndexPrepare indicates an expected call of OnIndexPrepare.
+func (mr *MockOnIndexSeriesMockRecorder) OnIndexPrepare(blockStart interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnIndexPrepare", reflect.TypeOf((*MockOnIndexSeries)(nil).OnIndexPrepare), blockStart)
+}
+
+// OnIndexSuccess mocks base method.
+func (m *MockOnIndexSeries) OnIndexSuccess(blockStart time.UnixNano) {
+	m.ctrl.T.Helper()
+	m.ctrl.Call(m, "OnIndexSuccess", blockStart)
+}
+
+// OnIndexSuccess indicates an expected call of OnIndexSuccess.
+func (mr *MockOnIndexSeriesMockRecorder) OnIndexSuccess(blockStart interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnIndexSuccess", reflect.TypeOf((*MockOnIndexSeries)(nil).OnIndexSuccess), blockStart)
+}
+
+// RelookupAndCheckIsEmpty mocks base method.
+func (m *MockOnIndexSeries) RelookupAndCheckIsEmpty() (bool, bool) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "RelookupAndCheckIsEmpty")
+	ret0, _ := ret[0].(bool)
+	ret1, _ := ret[1].(bool)
+	return ret0, ret1
+}
+
+// RelookupAndCheckIsEmpty indicates an expected call of RelookupAndCheckIsEmpty.
+func (mr *MockOnIndexSeriesMockRecorder) RelookupAndCheckIsEmpty() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RelookupAndCheckIsEmpty", reflect.TypeOf((*MockOnIndexSeries)(nil).RelookupAndCheckIsEmpty))
 }
